feat: moved big components into their own respective component files
