+++ apps/tenant-management-webapp/src/app/pages/admin/services/status/applications/StatusBar.ts
